Move test to test_human_eval.py
